requirements.txt
Django~=3.2.5
djangorestframework~=3.12.4
django-crispy-forms~=1.12.0
Pillow~=8.3.1
psycopg2-binary>=2.8
celery==5.1.2
redis==3.5.3